derive GIMP_TYPE_MEMSIZE from G_TYPE_ULONG. Should probably be even
2002-05-28  Sven Neumann  <sven@gimp.org>

	* app/config/gimpconfig-types.c: derive GIMP_TYPE_MEMSIZE from
	G_TYPE_ULONG. Should probably be even G_TYPE_UINT64 but we use
	strtol which can only handles unsigned long int.

	* app/config/gimpbaseconfig.[ch]
	* app/config/gimpguiconfig.[ch]: changed accordingly.

	* app/config/test-config.c: use gimp_config_serialize_value() to
	dump changed values to stdout.

+++ app/config/gimpbaseconfig.h
@@ -43,7 +43,7 @@ struct _GimpBaseConfig
   gchar    *swap_path;
   gboolean  stingy_memory_use;
   guint     num_processors;
-  guint     tile_cache_size;
+  gulong    tile_cache_size;
 };
 
 struct _GimpBaseConfigClass
